feat: add betting hook and header component
